Wristband Software V1.0-e
Act-on-the-go concept driven menu navigation system, currently support short press to switch page in all cases, long press(more than 900 millisecond), and one-key emergency request(more than 5500 millisecond)
Utilizing already created Two-way com channel, the wristband for the first time actively push message to the server to request for emergency
Minor version bump
This commit is the result of fixing or re-considering an issue about user interaction found when doing the first attempt to code the long-press function. So it's basically the second iteration. The issue is described in commit 92edce7