Wristband Software V1.1
Major Version Upgrade v1.0-f to v1.1
The user profile is no longer stored locally on the wristband, it is now dynamically allocated to the wristband upon boot by the server.
The server assign profile based on Hardware UUID(unique id) of each wristband. This is a very important step ahead for the entire project as it is no longer required to reprogram the wristband each time user/facility allocation changes, all job can be done on the server.
Because of that, now the time is also announced by the server to the wristband, after boot up, the RTC in the wristband keep track of the time
Will sync Server config in next commit