Code Re-organization && Bug fix
Re-organize the code to move UI section and not major code segment out of main code.  To keep main code clean and easy to work on.
Fix minor known bug.

Note: This version pass compile but is not tested on hardware. Use previous commit as rollback point if necessary. This code functionally should be same as previous commit.